pyproject.toml
[tool.poetry]
name = "nonebot_plugin_handle"
version = "0.1.6"
description = "Nonebot2 汉字Wordle 插件"
authors = ["meetwq <meetwq@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/noneplugin/nonebot-plugin-handle"
repository = "https://github.com/noneplugin/nonebot-plugin-handle"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0-beta.4"
nonebot-adapter-onebot = "^2.0.0-beta.1"
Pillow = "^9.0.0"
pypinyin = ">=0.40.0"

[tool.poetry.dev-dependencies]
nb-cli = "^0.6.0"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
